# docs/api.yml
openapi: 3.0.0
info:
  title: HospitalityPlatform API
  version: 1.0.0
  description: >-
    The API documentation for HospitalityPlatform, a cross-platform customer
    service and hotel management platform. <br><br>Please note that all API
    documentation will be done with a local instance of Swagger, will use
    OpenAPI 3.0.0, and [can be found on our
    GitHub](https://github.com/COP4331-Group-19/HospitalityPlatform).
  x-logo:
    url: ''
servers:
  - url: https://hospitalityplatform.herokuapp.com/api
    description: Production
  - url: https://virtserver.swaggerhub.com/jeffreydivi/HospitalityPlatform/1.0.0
    description: SwaggerHub API Auto Mocking
paths:
...
  /account/login/:
    post:
      security: []
      summary: Create a session.
      requestBody:
        description: Account credentials
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
        required: true
      tags:
        - Account
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewJWT'
          description: Return the currently logged-in user's account information.
      operationId: post_account_login
      description: Given a username and password, returns a JWT for the session.
...
components:
  schemas:
...
    NewJWT:
      required:
        - token
      type: object
      properties:
        token:
          description: The Bearer token
          type: string
          example: Bearer V2UncmUgbm8gc3RyYW5nZXJzIHRvIGxvdmUuIFlvdSBub3cgdGhlIHJ1bGVzIGFuZCBzbyBkbyBJLg==
    Credentials:
      required:
        - username
        - password
      type: object
      properties:
        username:
          description: A username.
          type: string
          example: divi
        password:
          description: A password.
          type: string
          example: CorrectHorseBatteryStaple
...
  securitySchemes:
    Bearer:
      scheme: bearer
      type: http
...
security:
  - Bearer: []
...